// src/theme/shared/types.ts
import { InvitationDataUser, InvitationEvent, InvitationGift, GuestBook } from '@/types/interface';

export interface CoverClasses {
  container?: string;
  overlay?: string;
  image?: string;
  gradient?: string;
  content?: string;
  title?: string;
  guestName?: string;
  guestLabel?: string;
  button?: string;
  buttonIcon?: string;
  loading?: string;
}

export interface IntroductionClasses {
  container?: string;
  label?: string;
  names?: string;
  date?: string;
  sectionTitle?: string;
  countdownContainer?: string;
  countdownBox?: string;
  countdownValue?: string;
  countdownLabel?: string;
}

export interface CoupleClasses {
  container?: string;
  marquee?: string;
  marqueeImage?: string;
  marqueeOverlayLeft?: string;
  marqueeOverlayRight?: string;
  groomSection?: string;
  groomPhoto?: string;
  groomRole?: string;
  groomName?: string;
  groomParent?: string;
  groomInstagram?: string;
  brideSection?: string;
  bridePhoto?: string;
  brideRole?: string;
  brideName?: string;
  brideParent?: string;
  brideInstagram?: string;
  divider?: string;
  dividerIcon?: string;
}

export interface QuotesClasses {
  container?: string;
  card?: string;
  verse?: string;
  reference?: string;
}

export interface EventClasses {
  container?: string;
  title?: string;
  subtitle?: string;
  titleDivider?: string;
  card?: string;
  cardTitle?: string;
  iconBox?: string;
  icon?: string;
  label?: string;
  value?: string;
  mapsButton?: string;
  mapsIcon?: string;
}

export interface GalleryClasses {
  container?: string;
  title?: string;
  grid?: string;
  image?: string;
  imageOverlay?: string;
  imageNumber?: string;
  lightboxOverlay?: string;
  lightboxClose?: string;
  lightboxCounter?: string;
  lightboxImage?: string;
  lightboxNav?: string;
}

export interface GiftClasses {
  container?: string;
  title?: string;
  subtitle?: string;
  titleDivider?: string;
  tabsContainer?: string;
  tabButton?: string;
  tabActive?: string;
  tabInactive?: string;
  tabIcon?: string;
  contentCard?: string;
  addressContainer?: string;
  addressIcon?: string;
  addressText?: string;
  bankIcon?: string;
  bankLabel?: string;
  bankDropdown?: string;
  bankDropdownMenu?: string;
  bankAccountCard?: string;
  walletIcon?: string;
  walletLabel?: string;
  walletDropdown?: string;
  walletDropdownMenu?: string;
  walletAddressCard?: string;
  copyButton?: string;
  copyIcon?: string;
  emptyState?: string;
}

export interface RSVPClasses {
  fab?: string;
  fabIcon?: string;
  fabBadge?: string;
  modalOverlay?: string;
  modalContainer?: string;
  modalHeader?: string;
  modalHeaderIcon?: string;
  modalTitle?: string;
  modalSubtitle?: string;
  modalClose?: string;
  chatContainer?: string;
  messageBubble?: string;
  messageAvatar?: string;
  messageName?: string;
  messageText?: string;
  messageTime?: string;
  dateSeparator?: string;
  inputContainer?: string;
  inputField?: string;
  sendButton?: string;
  sendIcon?: string;
  emptyState?: string;
}

export interface ThanksClasses {
  container?: string;
  content?: string;
  label?: string;
  names?: string;
  message?: string;
  brandingContainer?: string;
  brandingLabel?: string;
  brandingLogo?: string;
  socialContainer?: string;
  socialLink?: string;
  socialIcon?: string;
  copyright?: string;
  gradient?: string;
}

export interface LoveStoryClasses {
  container?: string;
  bgOrb1?: string;
  bgOrb2?: string;
  scrollContainer?: string;
  innerContainer?: string;
  timelineLine?: string;
  svgPath?: string;
  momentsContainer?: string;
  momentCard?: string;
  momentIcon?: string;
  momentTitle?: string;
  momentDate?: string;
  momentDateDot?: string;
  momentDescription?: string;
  momentGlow?: string;
  momentDot?: string;
  endingContainer?: string;
  endingText?: string;
  endingIcon?: string;
  scrollIndicator?: string;
  scrollLabel?: string;
  scrollArrow?: string;
}